ds_snowflake/requirements.txt
cnvrg
snowflake-connector-python
